requirement.txt
asgiref==3.3.4
certifi==2021.10.8
charset-normalizer==2.0.12
distlib==0.3.4
djangorestframework==3.13.1
filelock==3.7.1
idna==3.3
mysqlclient==2.1.0
Pillow==9.1.0
platformdirs==2.5.2
PyMySQL==1.0.2
pytz==2021.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.9
virtualenv==20.15.1
